+++ saleor/graphql/discount/mutations/sale/sale_remove_catalogues.py
@@ -5,9 +5,11 @@ from .....discount import models
 from .....discount.utils import fetch_catalogue_info
 from .....graphql.channel import ChannelContext
 from .....permission.enums import DiscountPermissions
+from .....webhook.event_types import WebhookEventAsyncType
 from ....core import ResolveInfo
 from ....core.doc_category import DOC_CATEGORY_DISCOUNTS
 from ....core.types import DiscountError
+from ....core.utils import WebhookEventInfo
 from ....plugins.dataloaders import get_plugin_manager_promise
 from ...types import Sale
 from ..utils import convert_catalogue_info_to_global_ids
@@ -21,6 +23,12 @@ class SaleRemoveCatalogues(SaleBaseCatalogueMutation):
         permissions = (DiscountPermissions.MANAGE_DISCOUNTS,)
         error_type_class = DiscountError
         error_type_field = "discount_errors"
+        webhook_events_info = [
+            WebhookEventInfo(
+                type=WebhookEventAsyncType.SALE_UPDATED,
+                description="A sale was updated.",
+            ),
+        ]
 
     @classmethod
     def perform_mutation(  # type: ignore[override]