+++ saleor/graphql/payment/mutations/transaction/transaction_request_action.py
@@ -23,7 +23,6 @@ from ....core.mutations import BaseMutation
 from ....core.scalars import PositiveDecimal
 from ....core.types import common as common_types
 from ....plugins.dataloaders import get_plugin_manager_promise
-from ....utils import get_user_or_app_from_context
 from ...enums import TransactionActionEnum
 from ...types import TransactionItem
 from .utils import get_transaction_item
@@ -59,17 +58,6 @@ class TransactionRequestAction(BaseMutation):
         doc_category = DOC_CATEGORY_PAYMENTS
         error_type_class = common_types.TransactionRequestActionError
         permissions = (PaymentPermissions.HANDLE_PAYMENTS,)
-
-    @classmethod
-    def check_permissions(cls, context, permissions=None, **data):
-        required_permissions = permissions or cls._meta.permissions
-        requestor = get_user_or_app_from_context(context)
-        for required_permission in required_permissions:
-            # We want to allow to call this mutation for requestor with one of following
-            # permission: manage_orders, handle_payments
-            if requestor and requestor.has_perm(required_permission):
-                return True
-        return False
 
     @classmethod
     def handle_transaction_action(